manifest: track some of our forks and PE stuff
Because of the way that our device trees work, we get to choose
between forking a lot of stuff from PixelOS and adapting it so that
it all works nicely with our trees, or we can just adapt our trees
a little and make PixelOS itself fetch stuff from PixelExperience
which is what our trees are supposed to work with, but we want them
to work with PixelOS, so let's get to that.

We also have to track some of our own forks for some stuff, such as
wpa_supplicant_8.

This entire manifest is focused towards making PixelOS work with
our device trees for the Poco X3/X3 NFC from PixelExperience, it is NOT
meant to be used anywhere else.